+++ server/services/caldav/api/caldav.controller.js
@@ -1,19 +1,18 @@
 module.exports = function CalDAVController(caldavHandler) {
-    /**
-     * @api {get} /api/v1/service/caldav/sync Start caldav sync for a user
-     * @apiName Sync
-     * @apiGroup CalDAV
-     */
-    async function sync(req, res) {
-      const bridges = await caldavHandler.sync('fc055c6b-4fb3-4127-820c-a05c45f72084');
-      res.json(bridges);
-    }
-  
-    return {
-      'get /api/v1/service/caldav/sync': {
-        authenticated: true,
-        controller: sync,
-      },
-    };
+  /**
+   * @api {get} /api/v1/service/caldav/sync Start caldav sync for a user
+   * @apiName Sync
+   * @apiGroup CalDAV
+   */
+  async function sync(req, res) {
+    const events = await caldavHandler.sync(req.user.id);
+    res.json(events);
+  }
+
+  return {
+    'get /api/v1/service/caldav/sync': {
+      authenticated: true,
+      controller: sync,
+    },
   };
-  
+};
