+++ server/services/caldav/index.js
@@ -1,10 +1,13 @@
 const logger = require('../../utils/logger');
 const CalDAVHandler = require('./lib');
+const CalDAVController = require('./api/caldav.controller');
 
 module.exports = function CalDAVService(gladys, serviceId) {
   const ical = require('ical');
   const dav = require('dav');
   const moment = require('moment');
+
+  const calDavHandler = new CalDAVHandler(gladys, serviceId, ical, dav, moment);
 
   /**
    * @public
@@ -29,6 +32,9 @@ module.exports = function CalDAVService(gladys, serviceId) {
   return Object.freeze({
     start,
     stop,
-    calendar: new CalDAVHandler(gladys, serviceId, ical, dav, moment),
+    calendar: {
+      sync: calDavHandler.sync,
+    },
+    controllers: CalDAVController(calDavHandler),
   });
 };
