--- server/services/caldav/lib/calendar/calendar.queries.js
@@ -1,3 +1,0 @@
-module.exports = {
-    getUsersQuery: `SELECT user as id FROM paramuser WHERE name = ?;`
-};
